+++ src/index.ts
@@ -17,6 +17,7 @@ export { InviteUserService } from './auth/invite-user.service';
 export { AuthService, hashPassword } from './auth/auth.service';
 export { UserService } from './auth/user.service';
 export { AUTH_LISTENER, AuthListener } from './auth/auth.listener';
+export { AUTH_CALLBACKS, AuthCallbacks } from './auth/auth.callbacks';
 export { Configuration, CONFIGURATION } from './configuration';
 export { SearchFields, SearchService } from './search/search.service';
 export { SearchableRepository } from './search/searchable.repository';