+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## 5.2.0 (2020-04-08)
+ - Optional auth callbacks to build user roles and build additional user properties (backwards-compatible ... not required)
+ - To utilise create an `@Injectable` class that implements `AuthCallbacks` and define one or more methods
+
 ## 5.1.2 (2020-01-31)
 
 - Fix typings to support multiple query predicates for arrays of primitives and union types.